Security changes

+++ retail-checkout-counter/src/main/java/org/mediaocean/rest/retail_checkout_counter/model/CategoryType.java
@@ -6,7 +6,7 @@ public enum CategoryType {
 
 	private double value;
 
-	private CategoryType(int value) {
+	private CategoryType(double value) {
 		this.value = value;
 	}
